Refactor `libcnb-data::buildpack`
Since at 500 lines (even before we finish adding tests), it was getting
a bit unwieldy.

The public API has not changed, aside from `BuildpackTomlError` having
been split into `BuildpackApiError` and `StackError`.

Future PRs will add additional unit tests to the split out types, and clean up
the TOML integration tests in the root of the `buildpack` module.
